% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dichotomous}
\alias{dichotomous}
\title{Estimates the parameters of a test}
\usage{
dichotomous(RData, dim, model, EMepsilon, theta, weights)
}
\arguments{
\item{RData}{Input dataset.}

\item{dim}{Model Dimension}

\item{model}{1 2 or 3PL}

\item{EMepsilon}{Convergence value for the algorithm}

\item{theta}{Quadrature Points}

\item{weights}{Quadrature Points Weights}
}
\description{
Estimates the parameters of a test
}

