% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpc.R
\name{eap}
\alias{eap}
\title{Latent Trait Estimation with EAP for gpc model}
\usage{
eap(pt.cuad, w.cuad, betas, pats, ncatg, dat.dic)
}
\arguments{
\item{pt.cuad}{TODO}

\item{w.cuad}{TODO}

\item{betas}{TODO}

\item{pats}{TODO}

\item{ncatg}{TODO}

\item{dat.dic}{TODO}
}
\description{
Estimates latent traits for gpc model.
}

