% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpc.R
\name{eap_gpc}
\alias{eap_gpc}
\title{Latent Trait Estimation with EAP for gerenalized parcial credit model (gpc) model}
\usage{
eap_gpc(estim)
}
\arguments{
\item{estim}{Output object from the estim_gpc function.}
}
\description{
Estimates latent traits for gerenalized parcial credit model (gpc) model.
Just for unidimensional models.
}
\examples{
\dontrun{
data = simulate_polytomous()$data
estim=estim_gpc(datos = datos)
eap_gpc = eap_gpc(estim)
plot(density(eap_gpc))
}
}

