% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpc.R
\name{estim_gpc}
\alias{estim_gpc}
\title{Estimation gpc model}
\usage{
estim_gpc(data)
}
\arguments{
\item{data}{The matrix containing the answers of tested individuals}
}
\description{
Estimates the test parameters according to the gerenalized parcial credit model (gpc) model.
This model is unidimensional only.
}
\examples{
\dontrun{
data = simulate_polytomous$data
estim=estim_gpc(data = datos)
}
}

