% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpc.R
\name{estim_grm}
\alias{estim_grm}
\title{Estimation gpc model}
\usage{
estim_grm(data)
}
\arguments{
\item{data}{The matrix containing the answers of tested individuals}
}
\description{
Estimates the test parameters according to gpc model, you only need
data matrix for estimation. This model is unidimensional only.
}

