% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inivalsMultiPoly.R
\name{inivals_MultiPoly}
\alias{inivals_MultiPoly}
\title{Initial values for multidimensional polytomous data}
\usage{
inivals_MultiPoly(data_poly, size.cluster, verbose = F)
}
\arguments{
\item{data_poly}{Polytomous data to get initial values.}

\item{size.cluster}{A vector with dimensionality test.}

\item{verbose}{True for get information about process in runtime. False in otherwise. False by default.}
}
\description{
Get initial values for polytomous data according dataset categories and
size cluster for dimensionality. It has no 3PL model. Then, no initial values for c parameter.
}

