% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inivalsMultiUni.R
\name{inivals_MultiUni}
\alias{inivals_MultiUni}
\title{Initial Values for one or more dimention models}
\usage{
inivals_MultiUni(data, size.cluster, model = "2PL",
  find.restrictions = FALSE, verbose = FALSE)
}
\description{
Find initial values for any data and model
}

