% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemfitbayesian.R
\name{itemfit.bayesian}
\alias{itemfit.bayesian}
\title{Parameter estimation of a test}
\usage{
itemfit.bayesian(data, dim, model = "2PL", EMepsilon = 1e-04,
  clusters = NULL, quad_tech = NULL, quad_points = NULL,
  individual_weights = as.integer(c()), initial_values = NULL,
  verbose = TRUE, save_time = TRUE)
}
\arguments{
\item{data}{The matrix containing the answers of tested individuals}

\item{dim}{The dimensionality of the test}

\item{model}{"1PL", "2PL" or "3PL"}

\item{EMepsilon}{Convergence value to determine the accuracy of the test}

\item{clusters}{A vector with cluster per dimension}

\item{quad_tech}{A string with technique. "Gaussian" for Gaussian quadrature 
or "QMCEM" for Quasi-Monte Carlo quadrature}

\item{quad_points}{Amount of quadrature points. If quadratura_technique is "Gaussian". It can be NULL}

\item{individual_weights}{A vector with Weights of the quadrature points.}

\item{initial_values}{A matrix with initial values for estimation process. Be sure about
dimension, model and consistency with data.}

\item{verbose}{True for get information about estimation process in runtime. False in otherwise.}

\item{save_time}{True for save estimation time. False otherwise.}
}
\description{
Estimates the test parameters according to the Multidimensional Item Response Theory with
bayesian adjust in dichotomous data
}

