% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentreg.R
\docType{package}
\name{LatentREGpp}
\alias{LatentREGpp}
\alias{LatentREGpp-package}
\title{LatentREGpp : Item Response Theory Implemented in R and Cpp}
\description{
latentregpp is a c++ implementation of the Multidimensional Item Respone Theory (MIRT)
cappable of performing parameter and traits estimations. It also provides a list of options to 
perform optiman analysis and provides usefull information about the obtained model.
}
\details{
\tabular{ll}{
Package: \tab LatentREGpp\cr
Type: \tab Package\cr
Version: \tab 0.0.5\cr
Date: \tab 2016-08-22\cr
License: \tab MIT + file LICENSE \cr
}
}
\section{Getting Started}{

Get started with the LatentREGpp package browsing the index of this documentation
if you need help the vignettes should be helpful.


The LatentREGpp package allows you to use the LatentREGpp methodology for simulating, analyzing and scoring tests \cr
You can browse the package vignettes to get started.
}
\author{
SICS Research Team
}
\keyword{IRT}
\keyword{MIRT}
\keyword{Psychometry}

