% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltraits.R
\name{ltraits}
\alias{ltraits}
\title{Latent traits estimation}
\usage{
ltraits(data, dim, model = "2PL", zetas = NULL, quad_tech = NULL,
  quad_points = NULL, init_traits = NULL, method = "MAP",
  by_individuals = TRUE, verbose = FALSE)
}

