% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpc.R
\name{map}
\alias{map}
\title{Latent Trait Estimation with MAP for gpc model}
\usage{
map(estim)
}
\arguments{
\item{estim}{Output object from the estim_grm function.}
}
\description{
Estimates latent traits for gpc model with MAP
data matrix for estimation. This model is unidimensional only.
}
\examples{
\dontrun{
data = simulate_polytomous()$data
estim=estim_grm(datos = datos)
map = map(estim)
}
}

