% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpc.R
\name{map_gpc}
\alias{map_gpc}
\title{Latent Trait Estimation with MAP for gerenalized parcial credit model (gpc)}
\usage{
map_gpc(estim)
}
\arguments{
\item{estim}{Output object from the estim_gpc function.}
}
\description{
Estimates latent traits for gerenalized parcial credit model (gpc) with MAP
data matrix for estimation. Just for unidimensional models.
}
\examples{
\dontrun{
data = simulate_polytomous()$data
estim=estim_gpc(datos = datos)
map_gpc = map_gpc(estim)
}
}

