% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multiTest_dico}
\alias{multiTest_dico}
\title{Estimates the parameters of a test}
\usage{
multiTest_dico(RData, dim, wd)
}
\arguments{
\item{RData}{Input dataset.}
}
\description{
Estimates the parameters of a test
}

