% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MulTRI.R
\docType{package}
\name{MulTRI}
\alias{MulTRI}
\alias{MulTRI-package}
\title{MulTRI : Item Response Theory Implemented in R and Cpp}
\description{
Mul-Tri is a c++ implementation of the Multidimensional Item Respone Theory (MIRT)
cappable of performing parameter and traits estimations. It also provides a list of options to 
perform optiman analysis and provides usefull information about the obtained model.
}
\details{
\tabular{ll}{
Package: \tab MulTRI\cr
Type: \tab Package\cr
Version: \tab 0.0.5\cr
Date: \tab 2016-08-22\cr
License: \tab MIT + file LICENSE \cr
}
}
\section{Getting Started}{

Get started with the MulTRI package browsing the index of this documentation
if you need help the vignettes should be helpful.


The IRTpp package allows you to use the MulTRI methodology for simulating, analyzing and scoring tests \cr
You can browse the package vignettes to get started.
}
\author{
SICS Research Team
}
\keyword{IRT}
\keyword{MIRT}
\keyword{Psychometry}

