% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s_ss}
\alias{s_ss}
\title{Calculates the likelihoods.}
\usage{
s_ss(pr, nitems, G)
}
\arguments{
\item{pr}{Probability matrix.}

\item{nitems}{Number of items considered in the probability matrix.}

\item{G}{the group number}
}
\description{
Likelihood of the classics scores to calculate 
expected frequencies in Orlando statistical.
}
\keyword{internal}

