% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_dichotomous.R
\name{simulate_dichotomous}
\alias{simulate_dichotomous}
\title{Dichotomous data simulation}
\usage{
simulate_dichotomous(dim.data = 1, sample.size = 1000,
  size.cluster = c(20), seed_data)
}
\arguments{
\item{dim.data}{Data dimension}

\item{sample.size}{Size of the population that will be simulated}

\item{size.cluster}{Vector containing the number of items by dimension}

\item{seed_data}{Seed for simulation}
}
\description{
This function generates dichotomous test of an specified dimension and size. The items per dimension are specified in the size.cluster vector, and the individuals are specified in sample.size.
}

