% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_polytomous.R
\name{simulate_polytomous}
\alias{simulate_polytomous}
\title{data simulation}
\usage{
simulate_polytomous(dim.data = 1, sample.size = 1000,
  size.cluster = c(20), ncatgs = rep(4, 20), seed_data)
}

