% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iniVals.R
\name{traits_patt2data}
\alias{traits_patt2data}
\title{Returns a list with the trait of every individual}
\usage{
traits_patt2data(pattern, pattern.traits, data)
}
\arguments{
\item{pattern}{Is the list with all response patterns}

\item{pattern.traits}{Is the list with the trait for each pattern response}

\item{the}{data used to make the estimation of traits and paramters}
}
\description{
Returns a list with the trait of every individual
}

