% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{x2}
\alias{x2}
\title{Statistical x2.}
\usage{
x2(z, patterns, G, FUN)
}
\arguments{
\item{patterns, }{list with Patterns, frequencies and traits.}

\item{G, }{the number of groups.}

\item{FUN, }{It is the function with which the expected probability 
is calculated in each group}

\item{model, }{The implemented model ("1PL","2PL","3PL")}

\item{zita}{matrix of estimations of the parameters of the items (alphas, d's, guessing).}
}
\description{
Calculates the statistical x2.
}

