% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodness_of_fit.R
\name{x2_itemf}
\alias{x2_itemf}
\title{Statistical x2.}
\usage{
x2_itemf(zetas, patterns, G, FUN)
}
\arguments{
\item{zetas}{matrix of estimations of the parameters of the items (alphas, d's, guessing).}

\item{patterns}{list with Patterns, frequencies and traits.}

\item{G}{the number of groups, by default 10}

\item{FUN}{It is the function with which the expected probability, by default median 
is calculated in each group.}
}
\description{
Calculates the statistical x2.
}

