% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodness_of_fit.R
\name{z3_itemf}
\alias{z3_itemf}
\title{Z3item fit statistic}
\usage{
z3_itemf(data, zita, patterns)
}
\arguments{
\item{data}{a data frame or a matrix with the test.}

\item{zita}{Matrix which columns are the estimations of the parameters of the items (discrimination,d=-discrimnation*difficulty, guessing).}

\item{patterns}{list of patterns response (patterns$patterns), the frequency of each pattern (patterns$freq) and the latent traits (patterns$latent_traits).}
}
\description{
Calculates the values of statistical Z3 for items.
}
\references{
Fritz Drasgow, Michael V. Levine and Esther A. Williams (1985). Appropiateness measurement with polychotomous item response models and standarized indices.
}
\seealso{
\code{\link{z3_personf}}, \code{\link{orlando_itemf}}
}

