% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpc.R
\name{eap}
\alias{eap}
\title{Latent Trait Estimation with EAP for gpc model}
\usage{
eap(estim)
}
\arguments{
\item{estim}{Output object from the estim_grm function.}
}
\description{
Estimates latent traits for gpc model.
}
\examples{
\dontrun{
data = simulate_polytomous()$data
estim=estim_grm(datos = datos)
eap = eap(estim)
plot(density(eap))
}
}

