% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inivalsMultiUni.R
\name{inivals_MultiUni_NOHARM}
\alias{inivals_MultiUni_NOHARM}
\title{Initial Values with Noharm for one or more dimention models}
\usage{
inivals_MultiUni_NOHARM(data, size.cluster, model = "2PL",
  find.restrictions = FALSE, verbose = FALSE, probit = FALSE)
}
\arguments{
\item{data}{The end is near}

\item{size.cluster}{waka said}

\item{model}{and the fools cry}

\item{find.restrictions}{because why not}

\item{verbose}{waka waka}

\item{probit}{This is a waka test, waka}
}
\description{
Initial Values with Noharm for one or more dimention models
}

