% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentreg.R
\name{latentreg}
\alias{latentreg}
\title{Parameter estimation of a test}
\usage{
latentreg(data, dim, model = "2PL", EMepsilon = 1e-04, clusters = NULL,
  quadrature_technique = NULL, quadrature_points = NULL,
  individual_weights = as.integer(c()), initial_values = NULL,
  verbose = TRUE)
}
\arguments{
\item{data}{The matrix containing the answers of tested individuals}

\item{dim}{The Dimensionality of the test}

\item{model}{1PL, 2PL or 3PL}

\item{EMepsilon}{Convergence value to determine the accuracy of the test}

\item{clusters}{Clusters per dimention}

\item{quadrature_points}{Amount of quadrature points}

\item{individual_weights}{Weights of the quadrature points}

\item{initial_values}{Initial Values of the estimation}

\item{quadratura_technique}{Quasi-Monte Carlo or Gaussian}
}
\description{
Estimates the test parameters according to the Multidimensional Item Response Theory
}

