% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemfit.R
\docType{package}
\name{LatentREGpp}
\alias{LatentREGpp}
\alias{LatentREGpp-package}
\title{LatentREGpp : Item Response Theory Implemented in R and Cpp}
\description{
This Package is a c++ implementation of the Multidimensional Item Response Theory (MIRT) capable of performing parameter and traits estimations. It also provides a list of options to perform an optimal analysis and provides useful information about the obtained model.\cr
This package is a work of SICS Research Group, Universidad Nacional de Colombia.\cr
}
\details{
\tabular{ll}{
Package: \tab LatentREGpp\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2016-09-07\cr
License: \tab MIT + file LICENSE \cr
}
}
\section{Getting Started}{

Get started with the LatentREGpp package browsing the index of this documentation
if you need help the vignettes should be helpful.


The LatentREGpp package allows you to use the LatentREGpp methodology for simulating, analyzing and scoring tests \cr
You can browse the package vignettes to get started.
}

\section{Acknowledgment}{

This work was Supported by Colciencias Research Grant 0039-2013 and SICS Research Group, Universidad Nacional de Colombia.
}
\author{
Milder Hernandez Cagua <milderhc@gmail.com>

Jhonatan Javier Guzman del Rio <jhjguzmanri@unal.edu.co>

Camilo Antonio Suarez Bolanos <caasuarezbo@unal.edu.co>

Alvaro Mauricio Montenegro <ammontenegrod@unal.edu.co>

Sergio Paez Moncaleano <spaezm@unal.edu.co>

Juan David Cortes Castillo <jdcortesc@unal.edu.co>

Campo Elias Pardo Turriago <cepardot@unal.edu.co>

Luisa Fernanda Parra Arboleda <lfparraar@unal.edu.co>

Emilio Pablo Berdugo Camacho <epberdugoc@unal.edu.co>
}
\keyword{IRT}
\keyword{MIRT}
\keyword{Psychometry}

