% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentregpp.R
\docType{package}
\name{latentregpp}
\alias{latentregpp}
\alias{latentregpp-package}
\title{latentregpp : Item Response Theory Implemented in R and Cpp}
\usage{
latentregpp(data, dim, model = "2PL", EMepsilon = 1e-04, clusters = NULL,
  quadrature_technique = NULL, quadrature_points = NULL,
  individual_weights = as.integer(c()), initial_values = NULL,
  verbose = TRUE)
}
\arguments{
\item{data}{The matrix containing the answers of tested individuals}

\item{dim}{The Dimensionality of the test}

\item{model}{1PL, 2PL or 3PL}

\item{EMepsilon}{Convergence value to determine the accuracy of the test}

\item{clusters}{Clusters per dimention}

\item{quadrature_points}{Amount of quadrature points}

\item{individual_weights}{Weights of the quadrature points}

\item{initial_values}{Initial Values of the estimation}

\item{quadratura_technique}{Quasi-Monte Carlo or Gaussian}
}
\description{
latentregpp is a c++ implementation of the Multidimensional Item Respone Theory (MIRT)
cappable of performing parameter and traits estimations. It also provides a list of options to 
perform optiman analysis and provides usefull information about the obtained model.

Estimates the test parameters according to the Multidimensional Item Response Theory
}
\details{
\tabular{ll}{
Package: \tab latentregpp\cr
Type: \tab Package\cr
Version: \tab 0.0.5\cr
Date: \tab 2016-08-22\cr
License: \tab MIT + file LICENSE \cr
}
}
\section{Getting Started}{

Get started with the latentregpp package browsing the index of this documentation
if you need help the vignettes should be helpful.


The latentregpp package allows you to use the latentregpp methodology for simulating, analyzing and scoring tests \cr
You can browse the package vignettes to get started.
}
\author{
SICS Research Team
}
\keyword{IRT}
\keyword{MIRT}
\keyword{Psychometry}

