% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpc.R
\name{map}
\alias{map}
\title{Latent Trait Estimation with MAP for gpc model}
\usage{
map(betas, nitems, ncatg)
}
\arguments{
\item{betas}{TODO}

\item{nitems}{TODO}

\item{ncatg}{TODO}
}
\description{
Estimates latent traits for gpc model with MAP
data matrix for estimation. This model is unidimensional only.
}

