% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MulTRI.R
\name{multri}
\alias{multri}
\title{Parameter estimation of a test}
\usage{
multri(data, dim, model = "2PL", EMepsilon = 1e-04, clusters = NULL,
  quadrature_technique = NULL, quadrature_points = NULL,
  indivudual_weights = NULL, initial_values = NULL)
}
\arguments{
\item{data}{The matrix containing the answers of tested individuals}

\item{dim}{The Dimensionality of the test}

\item{model}{1PL, 2PL or 3PL}

\item{EMepsilon}{Convergence value to determine the accuracy of the test}

\item{clusters}{Clusters per dimention}

\item{quadrature_points}{Amount of quadrature points}

\item{initial_values}{Initial Values of the estimation}

\item{quadratura_technique}{Quasi-Monte Carlo or Gaussian}

\item{individual_weights}{Weights of the quadrature points}
}
\description{
Estimates the test parameters according to the Multidimensional Item Response Theory
}

