% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_dichotomous.R
\name{simulate_dichotomous}
\alias{simulate_dichotomous}
\title{Dichotomous data simulation}
\usage{
simulate_dichotomous(dim.data = 1, sample.size = 1000,
  size.cluster = c(20), seed_data = 5000L, seed_item = 1000,
  model = "2PL")
}
\arguments{
\item{dim.data}{Data dimension}

\item{sample.size}{Size of the population to be simulated}

\item{size.cluster}{Vector containing the number of items per dimension}

\item{seed_data}{Seed for simulation of the data}

\item{seed_item}{Seed for simulation of the item parameters}

\item{model}{the model used to generate the test. It can have values of "1PL", "2PL" or "3PL".}
}
\value{
list with the simulated data, the information of the item parameters and
the individual parameters
}
\description{
This function generates dichotomous test
}
\examples{

#Example 1: see by default options
simm=simulate_dichotomous()
}

